Fix image rendering in local environment
Dynamix path parameters and nested dynamic routes are not working
in Nuxt 4 development mode. So this commit changes the code to use API
endpoints with query parameters which works in local development environments.
